+++ ccf-system/src/main/java/com/xyz/modules/biz/service/impl/ManageleadresponsInfoServiceImpl.java
@@ -1,21 +1,21 @@
 package com.xyz.modules.biz.service.impl;
 
-import cn.hutool.db.sql.Direction;
+
+import com.xyz.exception.BadRequestException;
 import com.xyz.modules.biz.domain.ManageleadresponsInfo;
-import com.xyz.modules.biz.service.dto.ManagecenterInfoDTO;
+import com.xyz.modules.security.security.JwtUser;
 import com.xyz.modules.system.domain.DictDetail;
-import com.xyz.modules.system.repository.DictDetailRepository;
 import com.xyz.modules.system.service.DictDetailService;
 import com.xyz.modules.system.util.DictEnum;
-import com.xyz.utils.ValidationUtil;
+import com.xyz.utils.*;
 import com.xyz.modules.biz.repository.ManageleadresponsInfoRepository;
 import com.xyz.modules.biz.service.ManageleadresponsInfoService;
 import com.xyz.modules.biz.service.dto.ManageleadresponsInfoDTO;
 import com.xyz.modules.biz.service.dto.ManageleadresponsInfoQueryCriteria;
 import com.xyz.modules.biz.service.mapper.ManageleadresponsInfoMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -26,8 +26,6 @@ import java.util.*;
 import cn.hutool.core.util.IdUtil;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import com.xyz.utils.PageUtil;
-import com.xyz.utils.QueryHelp;
 
 /**
 * @author xjh
@@ -46,6 +44,10 @@ public class ManageleadresponsInfoServiceImpl implements ManageleadresponsInfoSe
     @Autowired
     private DictDetailService dictDetailService;
 
+    @Autowired
+    @Qualifier("jwtUserDetailsService")
+    private UserDetailsService userDetailsService;
+
     @Override
     public Object queryAll(ManageleadresponsInfoQueryCriteria criteria, Pageable pageable){
         //Sort sort = new Sort(Sort.Direction.DESC,"createTime");
@@ -62,7 +64,8 @@ public class ManageleadresponsInfoServiceImpl implements ManageleadresponsInfoSe
         }
         Map map = new HashMap();
         map.put("content", manageleadresponsInfoList);
-        map.put("totalElements", page.getTotalPages());
+        map.put("totalElements", page.getTotalElements());
+        map.put("totalPages",page.getTotalPages());
         return map;
 
     }
@@ -74,33 +77,49 @@ public class ManageleadresponsInfoServiceImpl implements ManageleadresponsInfoSe
 
     @Override
     public ManageleadresponsInfoDTO findById(String id) {
-        Optional<ManageleadresponsInfo> ManageleadresponsInfo = ManageleadresponsInfoRepository.findById(id);
-        ValidationUtil.isNull(ManageleadresponsInfo,"ManageleadresponsInfo","id",id);
-        return ManageleadresponsInfoMapper.toDto(ManageleadresponsInfo.get());
+            if (StringUtils.isBlank(id)){
+                throw new BadRequestException("主键ID不能为空");
+            }
+            Optional<ManageleadresponsInfo> ManageleadresponsInfo = ManageleadresponsInfoRepository.findById(id);
+            ValidationUtil.isNull(ManageleadresponsInfo,"ManageleadresponsInfo","id",id);
+            return ManageleadresponsInfoMapper.toDto(ManageleadresponsInfo.get());
+
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public ManageleadresponsInfoDTO create(ManageleadresponsInfo resources) {
         resources.setId(IdUtil.simpleUUID());
-        resources.setCreateTime(new Timestamp(new Date().getTime()));
+        JwtUser u = (JwtUser) userDetailsService.loadUserByUsername(SecurityUtils.getUsername());
+        resources.setCreator(u.getId());
         return ManageleadresponsInfoMapper.toDto(ManageleadresponsInfoRepository.save(resources));
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void update(ManageleadresponsInfo resources) {
-        resources.setUpdateTime(new Timestamp(new Date().getTime()));
-        Optional<ManageleadresponsInfo> optionalManageleadresponsInfo = ManageleadresponsInfoRepository.findById(resources.getId());
-        ValidationUtil.isNull( optionalManageleadresponsInfo,"ManageleadresponsInfo","id",resources.getId());
-        ManageleadresponsInfo ManageleadresponsInfo = optionalManageleadresponsInfo.get();
-        ManageleadresponsInfo.copy(resources);
-        ManageleadresponsInfoRepository.save(ManageleadresponsInfo);
+            if (StringUtils.isBlank(resources.getId())){
+                throw new BadRequestException("主键ID不能为空");
+            }
+            Optional<ManageleadresponsInfo> optionalManageleadresponsInfo = ManageleadresponsInfoRepository.findById(resources.getId());
+            ValidationUtil.isNull( optionalManageleadresponsInfo,"ManageleadresponsInfo","id",resources.getId());
+            ManageleadresponsInfo ManageleadresponsInfo = optionalManageleadresponsInfo.get();
+            JwtUser u = (JwtUser) userDetailsService.loadUserByUsername(SecurityUtils.getUsername());
+            resources.setModifier(u.getId());
+            ManageleadresponsInfo.copy(resources);
+            ManageleadresponsInfoRepository.save(ManageleadresponsInfo);
+
+
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void delete(String id) {
+        if (StringUtils.isBlank(id)){
+            throw new BadRequestException("主键ID不能为空");
+        }
         ManageleadresponsInfoRepository.deleteById(id);
+
+
     }
 }
